+++ app/src/main/java/com/oliveira/meucaixa/ui/sales/CartItem.java
@@ -8,9 +8,9 @@ import com.oliveira.meucaixa.data.model.Product;
  */
 public class CartItem {
     private final Product product;
-    private int quantity;
+    private double quantity;
 
-    public CartItem(Product product, int quantity) {
+    public CartItem(Product product, double quantity) {
         this.product = product;
         this.quantity = quantity;
     }
@@ -19,11 +19,11 @@ public class CartItem {
         return product;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
